+++ weblate/checks/markup.py
@@ -6,6 +6,8 @@ from __future__ import annotations
 
 import re
 from collections import defaultdict
+from functools import cache, lru_cache
+from itertools import chain
 from typing import TYPE_CHECKING
 
 from django.core.exceptions import ValidationError
@@ -13,7 +15,10 @@ from django.core.validators import URLValidator
 from django.utils.functional import cached_property
 from django.utils.html import format_html_join
 from django.utils.safestring import mark_safe
-from django.utils.translation import gettext_lazy
+from django.utils.translation import gettext, gettext_lazy
+from docutils import utils
+from docutils.core import Publisher
+from docutils.nodes import Element, system_message
 
 from weblate.checks.base import MissingExtraDict, TargetCheck
 from weblate.utils.html import (
@@ -63,7 +68,33 @@ XML_ENTITY_MATCH = re.compile(
     re.VERBOSE,
 )
 
-RST_REF_MATCH = re.compile(r"((:[a-z:]+:)(?:`([^<`]+)`|`[^<`]+<([^<`]+)>`))")
+RST_REF_MATCH = re.compile(
+    r"(?:(?<=\W)|^)((:[a-z:]+:)(?:`([^<`]*[^<` ])`|`[^<`]+<([^<`]+)>`))(?=\W|$)"
+)
+RST_FOOTNOTE_MATCH = re.compile(r"(?:(?<=\W)|^)(\[#[^]]+\]_)(?=\W|$)")
+RST_LINK_MATCH = re.compile(r"(?:(?<=\W)|^)(`[^`]*[^` ]`_)(?=\W|$)")
+
+
+# These should be present in translation if present in source, but might be translated
+RST_TRANSLATABLE = {
+    ":guilabel:",
+    ":file:",
+    ":code:",
+    ":math:",
+    ":eq:",
+    ":abbr:",
+    ":dfn:",
+    ":menuselection:",
+    ":sub:",
+    ":sup:",
+    ":kbd:",
+    ":index:",
+}
+
+RST_ROLE_RE = [
+    re.compile(r"""Unknown interpreted text role "([^"]*)"\."""),
+    re.compile(r"""Interpreted text role "([^"]*)" not implemented\."""),
+]
 
 
 def strip_entities(text):
@@ -354,23 +385,32 @@ class SafeHTMLCheck(TargetCheck):
         return cleaned_target != target
 
 
-class RSTReferencesCheck(TargetCheck):
-    check_id = "rst-references"
-    name = gettext_lazy("Inconsistent reStructuredText references")
-    description = gettext_lazy(
-        "Inconsistent reStructuredText term references in the translated message."
-    )
+class RSTBaseCheck(TargetCheck):
     default_disabled = True
 
     def __init__(self) -> None:
         super().__init__()
         self.enable_string = "rst-text"
 
+
+class RSTReferencesCheck(RSTBaseCheck):
+    check_id = "rst-references"
+    name = gettext_lazy("Inconsistent reStructuredText references")
+    description = gettext_lazy(
+        "Inconsistent reStructuredText term references in the translated message."
+    )
+
     def extract_references(self, text: str) -> dict[str, str]:
-        return {
-            f"{role}{target or named_target}": text
+        result: dict[str, str] = {
+            role
+            if role in RST_TRANSLATABLE
+            else f"{role}{target or named_target}": text
             for text, role, target, named_target in RST_REF_MATCH.findall(text)
         }
+        result.update(
+            {footnote: footnote for footnote in RST_FOOTNOTE_MATCH.findall(text)}
+        )
+        return result
 
     def check_single(
         self, source: str, target: str, unit: Unit
@@ -381,12 +421,20 @@ class RSTReferencesCheck(TargetCheck):
         src_set = set(src_references.keys())
         tgt_set = set(tgt_references.keys())
 
+        errors: list[str] = []
+
         missing = src_set - tgt_set
         extra = tgt_set - src_set
-        if missing or extra:
+
+        src_links = len(RST_LINK_MATCH.findall(source))
+        tgt_links = len(RST_LINK_MATCH.findall(target))
+        if src_links != tgt_links:
+            errors.append(gettext("Inconsistent links in the translated message."))
+        if missing or extra or errors:
             return {
                 "missing": [src_references[item] for item in missing],
                 "extra": [tgt_references[item] for item in extra],
+                "errors": errors,
             }
         return False
 
@@ -416,5 +464,116 @@ class RSTReferencesCheck(TargetCheck):
     def check_highlight(self, source: str, unit: Unit):
         if self.should_skip(unit):
             return
-        for match in RST_REF_MATCH.finditer(source):
+        for match in chain(
+            RST_REF_MATCH.finditer(source), RST_FOOTNOTE_MATCH.finditer(source)
+        ):
             yield match.start(0), match.end(0), match.group(0)
+
+
+@cache
+def get_rst_publisher() -> Publisher:
+    publisher = Publisher(settings=None)
+    publisher.set_components("standalone", "restructuredtext", "null")
+    publisher.get_settings(halt_level=5)
+    return publisher
+
+
+@lru_cache(maxsize=512)
+def validate_rst_snippet(
+    snippet: str, source_tags: tuple[str] | None = None
+) -> tuple[list[str], list[str]]:
+    publisher = get_rst_publisher()
+    document = utils.new_document(None, publisher.settings)
+    document.reporter.stream = None
+
+    errors: list[str] = []
+    roles: list[str] = []
+
+    def error_collector(data: system_message) -> None:
+        """Save the error."""
+        message = Element.astext(data)
+        if message.startswith("Unknown target name:") and "`" not in message:
+            # Translating targets is okay, just catch obvious errors
+            return
+        if message.startswith(
+            (
+                # Duplicates Unknown interpreted in our case
+                "No role entry",
+                # Can not work on snippets
+                "Too many autonumbered footnote",
+                # Can not work on snippets
+                "Enumerated list start value not ordinal",
+            )
+        ):
+            return
+        for rst_role_re in RST_ROLE_RE:
+            if match := rst_role_re.match(message):
+                role = match.group(1)
+                roles.append(role)
+                if source_tags is not None and role in source_tags:
+                    # Skip if the role was found in the source
+                    return
+        errors.append(message)
+
+    document.reporter.attach_observer(error_collector)
+    publisher.reader.parser.parse(snippet, document)
+    transformer = document.transformer
+    transformer.populate_from_components(
+        (
+            publisher.source,
+            publisher.reader,
+            publisher.reader.parser,
+            publisher.writer,
+            publisher.destination,
+        )
+    )
+    while transformer.transforms:
+        if not transformer.sorted:
+            # Unsorted initially, and whenever a transform is added.
+            transformer.transforms.sort()
+            transformer.transforms.reverse()
+            transformer.sorted = 1
+        priority, transform_class, pending, kwargs = transformer.transforms.pop()
+        transform = transform_class(transformer.document, startnode=pending)
+        transform.apply(**kwargs)
+        transformer.applied.append((priority, transform_class, pending, kwargs))
+    return errors, roles
+
+
+class RSTSyntaxCheck(RSTBaseCheck):
+    check_id = "rst-syntax"
+    name = gettext_lazy("reStructuredText syntax error")
+    description = gettext_lazy("reStructuredText syntax error in the translation.")
+
+    def check_single(
+        self, source: str, target: str, unit: Unit
+    ) -> bool | MissingExtraDict:
+        _errors, source_tags = validate_rst_snippet(source)
+        errors, _target_tags = validate_rst_snippet(target, tuple(source_tags))
+
+        if errors:
+            return {"errors": errors}
+        return False
+
+    def get_description(self, check_obj: Check) -> StrOrPromise:
+        unit = check_obj.unit
+
+        errors: list[StrOrPromise] = []
+        results: MissingExtraDict = defaultdict(list)
+
+        # Merge plurals
+        for result in self.check_target_generator(
+            unit.get_source_plurals(), unit.get_target_plurals(), unit
+        ):
+            if isinstance(result, dict):
+                for key, value in result.items():
+                    results[key].extend(value)
+        if results:
+            errors.extend(self.format_result(results))
+        if errors:
+            return format_html_join(
+                mark_safe("<br />"),  # noqa: S308
+                "{}",
+                ((error,) for error in errors),
+            )
+        return super().get_description(check_obj)